+++ lib/act_as_fire_record_beta.rb
@@ -81,8 +81,7 @@ module ActAsFireRecordBeta
     end
 
     def first(limit = 1)
-      records = col.limit(limit)
-      limit == 1 ? records[0] : records
+      all.first(limit)
     end
 
     def count