+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Next
 - Remove `get_records` method. Now records can be retrieved without it.
+- Delegate to Array on method missing for Google::Cloud::Firestore::Query (e.g. `query.size`, `query.empty?`, etc).
 - Add `count` method to model class. (e.g. `Book.count`)
+
 ## v0.0.8 - 2022-11-24
 - `all` returns `Google::Cloud::Firestore::CollectionReference` object. `get_records` is required to fetch data from Firestore.